Update patch set 1
Patch Set 1:

-1,

Can we revert this patch? or fix the code. This patch broke autorelease and adds bad files into the repo such as impl/pom.xml~ which is a temp file.

The pom files reverts backwards to dependencies that are for Lithium and not the Beryllium release. For example binding-parent was moved from yangtools to mdsal. We fixed this in usecplugin but this patch reintroduces the issue :(

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0