Update patch set 1
Patch Set 1:

Thomas there is no need to abandon the patch!

Gerrit will update this patch if you keep the same Change-Id field. This is the correct workflow for Gerrit.

All you had to do is run "git commit --amend" and update the patch leaving the Change-Id field in tact.

Patch-set: 1